Give each memory adapter its own module.
The vendor classes lived in one providers.py file behind re-export stubs.
Moving them into the named modules, with shared answering helpers in common.py,
makes it easier to find and change a single provider.
